requirements.txt
# Front Facing Webpage
streamlit

# Database & Cloud
supabase
pandas

# Web Scraping
playwright
beautifulsoup4

# Utilities
python-dotenv
requests